requirements.txt
gdown
matplotlib
numba==0.49.1
numpy==1.18.3
icecream
scikit-image
albumentations
einops
linformer
timm
#--find-links https://download.pytorch.org/whl/torch_stable.html
#torch==1.7.1+cu110
#torchvision==0.8.2+cu110
#torchaudio==0.7.2
#tensorflow
wandb
adamp
git+https://github.com/fadel/pytorch_ema
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git
git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup
git+https://github.com/pabloppp/pytorch-tools
git+https://github.com/JunnYu/mish-cuda
adabelief-pytorch==0.2.0
hydra
mlflow
hydra-optuna-sweeper --upgrade
